# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "VASARI API",
    "description": "Documentation automatically generated by the <b>swagger-autogen</b> module."
  },
  "host": "localhost:3000",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/schedeRA": {
      "get": {
        "tags": [
          "Schede RA"
        ],
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "Schede RA"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "mt": {
                  "example": "any"
                },
                "co": {
                  "example": "any"
                },
                "cm": {
                  "example": "any"
                },
                "rs": {
                  "example": "any"
                },
                "au": {
                  "example": "any"
                },
                "rv": {
                  "example": "any"
                },
                "lc": {
                  "example": "any"
                },
                "og": {
                  "example": "any"
                },
                "cd": {
                  "example": "any"
                },
                "ub": {
                  "example": "any"
                },
                "dt": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/schedeRA/ogtn/{ogtn}": {
      "get": {
        "tags": [
          "Schede RA"
        ],
        "description": "",
        "parameters": [
          {
            "name": "ogtn",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/schedeRA/{id}": {
      "delete": {
        "tags": [
          "Schede RA"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "get": {
        "tags": [
          "Schede RA"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "patch": {
        "tags": [
          "Schede RA"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "mt": {
                  "example": "any"
                },
                "co": {
                  "example": "any"
                },
                "cm": {
                  "example": "any"
                },
                "rs": {
                  "example": "any"
                },
                "au": {
                  "example": "any"
                },
                "rv": {
                  "example": "any"
                },
                "lc": {
                  "example": "any"
                },
                "og": {
                  "example": "any"
                },
                "cd": {
                  "example": "any"
                },
                "ub": {
                  "example": "any"
                },
                "dt": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
...
}